server/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
google-genai
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pgvector
